Use the most up to date deep link URL for BankID
A different one is specified in
https://developer.android.com/training/package-visibility/declaring#intent-filter-signature
compared to what we had before

We suspect this meant that for some people the BankID app was not
properly found, so we did not show the right information on the screen
