+++ app/feature/feature-login/src/main/kotlin/com/hedvig/android/feature/login/swedishlogin/SwedishLoginPresenter.kt
@@ -220,7 +220,7 @@ internal sealed interface BankIdUiState {
     value class AutoStartToken(val token: String) {
       // The Uri which opens the BankId app while also passing in the right autoStartUrl
       val bankIdUri: Uri
-        get() = Uri.parse("bankid:///?autostarttoken=$token&redirect=null")
+        get() = Uri.parse("https://app.bankid.com/?autostarttoken=$token&redirect=null")
     }
 
     @JvmInline
